# docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "go实战",
    "title": "博客系统",
    "termsOfService": "https://github.com/xieyuming/go-app.git",
    "contact": {},
    "license": {},
    "version": "1.0"
  },
  "paths": {
    "/api/v1/tags": {
      "get": {
        "produces": [
          "application/json"
        ],
        "summary": "获取多个标签",
        "parameters": [
          {
            "maxLength": 100,
            "type": "string",
            "description": "标签名称",
            "name": "name",
            "in": "query"
          },
          {
            "enum": [
              0,
              1
            ],
            "type": "integer",
            "default": 1,
            "description": "状态",
            "name": "state",
            "in": "query"
          },
          {
            "type": "integer",
            "description": "页码",
            "name": "page",
            "in": "query"
          },
          {
            "type": "integer",
            "description": "每页数量",
            "name": "page_size",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "schema": {
              "$ref": "#/definitions/models.TagSwagger"
            }
          },
          "400": {
            "description": "请求错误",
            "schema": {
              "$ref": "#/definitions/ecode.Error"
            }
          },
          "500": {
            "description": "内部错误",
            "schema": {
              "$ref": "#/definitions/ecode.Error"
            }
          }
        }
      },
      "post": {
        "produces": [
          "application/json"
        ],
        "summary": "新增标签",
        "parameters": [
          {
            "maxLength": 100,
            "minLength": 3,
            "description": "标签名称",
            "name": "name",
            "in": "body",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "enum": [
              0,
              1
            ],
            "default": 1,
            "description": "状态",
            "name": "state",
            "in": "body",
            "schema": {
              "type": "integer"
            }
          },
          {
            "maxLength": 100,
            "minLength": 3,
            "description": "创建者",
            "name": "created_by",
            "in": "body",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "schema": {
              "$ref": "#/definitions/models.Tag"
            }
          },
          "400": {
            "description": "请求错误",
            "schema": {
              "$ref": "#/definitions/ecode.Error"
            }
          },
          "500": {
            "description": "内部错误",
            "schema": {
              "$ref": "#/definitions/ecode.Error"
            }
          }
        }
      }
    },
    "/api/v1/tags/{id}": {
      "put": {
        "produces": [
          "application/json"
        ],
        "summary": "更新标签",
        "parameters": [
          {
            "type": "integer",
            "description": "标签ID",
            "name": "id",
            "in": "path",
            "required": true
          },
          {
            "maxLength": 100,
            "minLength": 3,
            "description": "标签名称",
            "name": "name",
            "in": "body",
            "schema": {
              "type": "string"
            }
          },
          {
            "enum": [
              0,
              1
            ],
            "default": 1,
            "description": "状态",
            "name": "state",
            "in": "body",
            "schema": {
              "type": "integer"
            }
          },
          {
            "maxLength": 100,
            "minLength": 3,
            "description": "修改者",
            "name": "modified_by",
            "in": "body",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/models.Tag"
              }
            }
          },
          "400": {
            "description": "请求错误",
            "schema": {
              "$ref": "#/definitions/ecode.Error"
            }
          },
          "500": {
            "description": "内部错误",
            "schema": {
              "$ref": "#/definitions/ecode.Error"
            }
          }
        }
      },
      "delete": {
        "produces": [
          "application/json"
        ],
        "summary": "删除标签",
        "parameters": [
          {
            "type": "integer",
            "description": "标签ID",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "schema": {
              "type": "string"
            }
          },
          "400": {
            "description": "请求错误",
            "schema": {
              "$ref": "#/definitions/ecode.Error"
            }
          },
          "500": {
            "description": "内部错误",
            "schema": {
              "$ref": "#/definitions/ecode.Error"
            }
          }
        }
      }
    }
  },
  "definitions": {
    "app.Pager": {
      "type": "object",
      "properties": {
        "page": {
          "type": "integer"
        },
        "page_size": {
          "type": "integer"
        },
        "total_rows": {
          "type": "integer"
        }
      }
    },
    "ecode.Error": {
      "type": "object"
    },
    "models.Tag": {
      "type": "object",
      "properties": {
        "created_by": {
          "type": "string"
        },
        "created_date": {
          "type": "string"
        },
        "deleted_date": {
          "type": "string"
        },
        "id": {
          "type": "integer"
        },
        "is_del": {
          "type": "integer"
        },
        "name": {
          "type": "string"
        },
        "status": {
          "type": "string"
        },
        "updated_by": {
          "type": "string"
        },
        "updated_date": {
          "type": "string"
        }
      }
    },
    "models.TagSwagger": {
      "type": "object",
      "properties": {
        "list": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/models.Tag"
          }
        },
        "pager": {
          "type": "object",
          "$ref": "#/definitions/app.Pager"
        }
      }
    }
  }
}